Finished converting all int64_t and uint64_t to 32 bit versions to improve cross-platform compatibility later. I do not yet see any issues from this. Also the benchmark kernel code is now officially very messy due to tons of OS and architectural dependences and pound-defs :(

+++ src/LatencyWorker.cpp
@@ -50,7 +50,7 @@ LatencyWorker::LatencyWorker(
 		void* mem_array,
 		size_t len,
 	#ifdef USE_SIZE_BASED_BENCHMARKS
-		uint64_t passes_per_iteration,
+		uint32_t passes_per_iteration,
 	#endif
 		RandomFunction kernel_fptr,
 		RandomFunction kernel_dummy_fptr,
@@ -127,7 +127,7 @@ void LatencyWorker::run() {
 		std::cerr << "WARNING: Failed to boost scheduling priority. Perhaps running in Administrator mode would help." << std::endl;
 
 	//Prime memory
-	for (uint64_t i = 0; i < 4; i++) {
+	for (uint32_t i = 0; i < 4; i++) {
 		void* prime_start_address = mem_array; 
 		void* prime_end_address = reinterpret_cast<void*>(reinterpret_cast<uint8_t*>(mem_array) + len);
 		forwSequentialRead_Word64(prime_start_address, prime_end_address); //dependent reads on the memory, make sure caches are ready, coherence, etc...
